Split compilation
Moved the Display_Welcome_Message function to output the program initial
 message to the header file.

version 1.4.0

resoved : #14
